Subpackages added and corresponding classes moved to them. Code formatted (identions and empty lines were reviewed).
