+++ src/test/java/lesson01/loopspracticing/MaxMinAverageTests.java
@@ -1,4 +1,4 @@
-package firstsimpletasks;
+package lesson01.loopspracticing;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -8,7 +8,6 @@ public class MaxMinAverageTests {
     private final int[] arrayFrom0To9 = {0,1,2,3,4,5,6,7,8,9};
     private final int[] arrayZeroLength = {};
     private final int[] arrayWithNegativeValues = {-5,225,-1,0,25,-61};
-
 
     @Test
     public void testGettingMaxValue() {
